Move common similarity methods to utils file
